+++ lib/v2/bloomberg/index.js
@@ -1,5 +1,4 @@
-const { rootUrl, parseNewsList, parseArticle } = require('./utils');
-const asyncPool = require('tiny-async-pool');
+const { rootUrl, asyncPoolAll, parseNewsList, parseArticle } = require('./utils');
 const site_title_mapping = {
     '/': 'News',
     bpol: 'Politics',
@@ -13,14 +12,6 @@ const site_title_mapping = {
     equality: 'Equality',
     businessweek: 'Businessweek',
     citylab: 'CityLab',
-};
-
-const asyncPoolAll = async (...args) => {
-    const results = [];
-    for await (const result of asyncPool(...args)) {
-        results.push(result);
-    }
-    return results;
 };
 
 module.exports = async (ctx) => {
