+++ lib/v2/bloomberg/utils.js
@@ -1,5 +1,6 @@
 const cheerio = require('cheerio');
 const path = require('path');
+const asyncPool = require('tiny-async-pool');
 
 const { parseDate } = require('@/utils/parse-date');
 const got = require('@/utils/got');
@@ -388,8 +389,17 @@ const processBody = async (body_html, story_json) => {
     return $.html();
 };
 
+const asyncPoolAll = async (...args) => {
+    const results = [];
+    for await (const result of asyncPool(...args)) {
+        results.push(result);
+    }
+    return results;
+};
+
 module.exports = {
     rootUrl,
+    asyncPoolAll,
     parseNewsList,
     parseArticle,
 };
